Refactored forms to use model for the data.
