+++ NextGraphics.csproj
@@ -163,11 +163,11 @@
     <Compile Include="fileTree.Designer.cs">
       <DependentUpon>fileTree.cs</DependentUpon>
     </Compile>
-    <Compile Include="settingsPanel.cs">
+    <Compile Include="SettingsPanel.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="settingsPanel.Designer.cs">
-      <DependentUpon>settingsPanel.cs</DependentUpon>
+    <Compile Include="SettingsPanel.Designer.cs">
+      <DependentUpon>SettingsPanel.cs</DependentUpon>
     </Compile>
     <Compile Include="ImageSelect.cs">
       <SubType>Form</SubType>
@@ -187,11 +187,11 @@
     <Compile Include="Main.Designer.cs">
       <DependentUpon>Main.cs</DependentUpon>
     </Compile>
-    <Compile Include="Palette.cs">
+    <Compile Include="PaletteForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Palette.Designer.cs">
-      <DependentUpon>Palette.cs</DependentUpon>
+    <Compile Include="PaletteForm.Designer.cs">
+      <DependentUpon>PaletteForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Progress.cs">
@@ -228,8 +228,8 @@
     <EmbeddedResource Include="rebuild.resx">
       <DependentUpon>rebuild.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="settingsPanel.resx">
-      <DependentUpon>settingsPanel.cs</DependentUpon>
+    <EmbeddedResource Include="SettingsPanel.resx">
+      <DependentUpon>SettingsPanel.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="ImageSelect.resx">
       <DependentUpon>ImageSelect.cs</DependentUpon>
@@ -240,8 +240,8 @@
     <EmbeddedResource Include="Main.resx">
       <DependentUpon>Main.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Palette.resx">
-      <DependentUpon>Palette.cs</DependentUpon>
+    <EmbeddedResource Include="PaletteForm.resx">
+      <DependentUpon>PaletteForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Progress.resx">
       <DependentUpon>Progress.cs</DependentUpon>
